requirements.txt
imutils
opencv-python
python-dotenv
numpy